lambda_functions/requirements.txt
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pyjwt==2.8.0
passlib[bcrypt]==1.7.4
